+++ src/app/app.component.ts
@@ -47,7 +47,7 @@ export class MyApp {
           location: 'browser',
           //logging: ['error', 'query', 'schema'],
           synchronize: true,
-          //dropSchema: true,
+          dropSchema: true,
           entities: [
             Category,
             Product,
