+++ src/pages/create-category/create-category.html
@@ -1,7 +1,7 @@
 <ion-header>
 
   <ion-navbar>
-    <ion-title>create-category</ion-title>
+    <ion-title>Agregar nueva categoria</ion-title>
   </ion-navbar>
 
 </ion-header>
